add user info

+++ utils/localStorage.util.ts
@@ -1,17 +1,17 @@
 function setLocalStorageContent(key: string, value: string): void {
-  window.localStorage.setItem(key, value);
+  if (typeof window !== "undefined") window.localStorage.setItem(key, value);
 }
 
-function getItem(key: string): string | null {
-  return window.localStorage.getItem(key);
+function getItem(key: string): string | null | undefined {
+  if (typeof window !== "undefined") return window.localStorage.getItem(key);
 }
 
 function removeItem(key: string): void {
-  return window.localStorage.removeItem(key);
+  if (typeof window !== "undefined") return window.localStorage.removeItem(key);
 }
 
 function clearLocalStorageContent(): void {
-  window.localStorage.clear();
+  if (typeof window !== "undefined") window.localStorage.clear();
 }
 
 export {
